% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_2df}
\alias{fun_2df}
\title{conver data.frame-like obj into data.frame}
\usage{
fun_2df(input_df, index_col = 1)
}
\arguments{
\item{input_df}{data.frame-like obj, like data.table,tibble,etc}

\item{index_col}{which coloumn to regard as rownames}
}
\value{
df
}
\description{
conver data.frame-like obj into data.frame
}
