% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_diff.R
\name{fun_deg_cutoff}
\alias{fun_deg_cutoff}
\title{Deg cutoff function}
\usage{
fun_deg_cutoff(
  input_df,
  pval = "adj.P.Val",
  pval_cutoff = 0.01,
  logFC = "logFC",
  logFC_cutoff = 1
)
}
\arguments{
\item{input_df}{data}

\item{pval}{pval}

\item{pval_cutoff}{pval_cutoff}

\item{logFC}{logFC}

\item{logFC_cutoff}{logFC_cutoff}
}
\value{
add p_sig column into input_df
}
\description{
Deg cutoff function
}
