% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_diff.R
\name{fun_deg_tpm_limma}
\alias{fun_deg_tpm_limma}
\title{Differential expression(TPM/FPKM/RPKM) analysis for genes, the input genes should be log-transformed, Notice that it maybe not the optimal methods for differential expression analysis.}
\usage{
fun_deg_tpm_limma(input_anno, input_expr, contrast = c("High", "Low"))
}
\arguments{
\item{input_anno}{Input annotation, first column is group}

\item{input_expr}{the expression matrix}

\item{contrast}{contrast group}
}
\value{
data.frame of differential expression table
}
\description{
Differential expression(TPM/FPKM/RPKM) analysis for genes, the input genes should be log-transformed, Notice that it maybe not the optimal methods for differential expression analysis.
}
