% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_design_cox}
\alias{fun_design_cox}
\title{Feature selection methods}
\usage{
fun_design_cox(input_methods, notUsedForTraining = c("rf", "unicox"))
}
\arguments{
\item{input_methods}{cancidate methods}

\item{notUsedForTraining}{not used as train}
}
\value{
df
}
\description{
Feature selection methods
}
