% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{fun_df_combine_columns}
\alias{fun_df_combine_columns}
\title{combine data.frame}
\usage{
fun_df_combine_columns(
  input_df,
  input_variables,
  new_var = "others",
  anno_mark = ":",
  sep = ";"
)
}
\arguments{
\item{input_df}{data.frame}

\item{input_variables}{input variables to collapse}

\item{new_var}{new var name}

\item{anno_mark}{the mark separate the variable name and value}

\item{sep}{the mark separate the variables}
}
\value{
data.frame
}
\description{
combine data.frame
}
