% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_cutoff.R
\name{fun_df_cutoff_fisher}
\alias{fun_df_cutoff_fisher}
\title{fisher test}
\usage{
fun_df_cutoff_fisher(
  input_df,
  input_variables,
  input_group = "Group",
  output_variables_names = input_variables[1],
  return_grouped_df = T,
  fun_surv_cutoff_args = list(),
  input_pct = seq(0.25, 0.75, 0.01),
  fun_stat_fisher_by_column_args = list()
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_group}{input_group}

\item{output_variables_names}{outputnames}

\item{return_grouped_df}{return_grouped_df}

\item{fun_surv_cutoff_args}{fun_surv_cutoff_args}

\item{input_pct}{input_pct}

\item{fun_stat_fisher_by_column_args}{fun_stat_fisher_by_column_args}
}
\value{
data.frame,results or grouped df
}
\description{
fisher test
}
