% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{fun_df_filter_duplicated}
\alias{fun_df_filter_duplicated}
\title{Filter the gene expression dataframe by genes, the duplicated genes with larger expression values would be kept}
\usage{
fun_df_filter_duplicated(
  input_df,
  input_group = "gene_name",
  input_numeric_columns = NULL
)
}
\arguments{
\item{input_df}{data.frame of gene expression}

\item{input_group}{which column is gene column}

\item{input_numeric_columns}{the expression column, NULL means all numeric columns}
}
\value{
df
}
\description{
Filter the gene expression dataframe by genes, the duplicated genes with larger expression values would be kept
}
