% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_diff.R
\name{fun_dmr_champ}
\alias{fun_dmr_champ}
\title{Differential Methylation Region(DMR) using ChAMP}
\usage{
fun_dmr_champ(
  input_df,
  input_matrix,
  outfile,
  input_group = "Group",
  input_sample = "sample",
  input_batch = NULL,
  filterXY = F,
  arraytype = "450K",
  cores = 2,
  method = "BMIQ",
  adjPVal = 0.05,
  compare.group = c("cancer", "normal")
)
}
\arguments{
\item{input_df}{pd}

\item{input_matrix}{beta}

\item{outfile}{outfile}

\item{input_group}{group}

\item{input_sample}{sample column}

\item{input_batch}{input_batch}

\item{filterXY}{T or F}

\item{arraytype}{arraytype}

\item{cores}{cores}

\item{method}{method}

\item{adjPVal}{adjPVal}

\item{compare.group}{compare.group}
}
\value{
list with differntial expressed genes
}
\description{
Differential Methylation Region(DMR) using ChAMP
}
