% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich.R
\name{fun_enrich_gsea}
\alias{fun_enrich_gsea}
\title{GSEA analysis}
\usage{
fun_enrich_gsea(
  input_genes,
  input_gs,
  pval = 0.2,
  logFC = "logFC",
  gene = "ID",
  term2gene = c("term", "gene")
)
}
\arguments{
\item{input_genes}{data.frame contains gene}

\item{input_gs}{term2gene}

\item{pval}{pval cutoff}

\item{logFC}{logFC column}

\item{gene}{gene colum}

\item{term2gene}{columns specify term and gene location}
}
\value{
enrich results
}
\description{
GSEA analysis
}
