% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich.R
\name{fun_enrich_ora}
\alias{fun_enrich_ora}
\title{ORA enerichment using clusterprofiler}
\usage{
fun_enrich_ora(
  input_genes,
  input_gs,
  pval = 0.2,
  term2gene = c("term", "gene")
)
}
\arguments{
\item{input_genes}{genes for test}

\item{input_gs}{geneset data, must have term and gene column}

\item{pval}{pval cutoff for pvalue and qvalue}

\item{term2gene}{columns specify term and gene location}
}
\value{
enrichment results
}
\description{
ORA enerichment using clusterprofiler
}
