% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrich.R
\name{fun_enrich_patch}
\alias{fun_enrich_patch}
\title{GO, kegg and hallmark enrichment analysis for input_data,}
\usage{
fun_enrich_patch(input_df, input_methods = "ora", pval = 0.2)
}
\arguments{
\item{input_df}{data.frame for gsea or gene list for ora}

\item{input_methods}{ora or gse}

\item{pval}{pval cutoff for pvalue and qvalue}
}
\value{
list of different enrichment obj
}
\description{
GO, kegg and hallmark enrichment analysis for input_data,
}
