% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_sig.R
\name{fun_extend_imvigor_mut}
\alias{fun_extend_imvigor_mut}
\title{Merge gene mutation status into clinical data}
\usage{
fun_extend_imvigor_mut(input_df, input_variables)
}
\arguments{
\item{input_df}{imvigor clinical data}

\item{input_variables}{genes}
}
\value{
data.frame, the merged df with genes mutation status
}
\description{
Merge gene mutation status into clinical data
}
