% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_sig.R
\name{fun_extend_sig_emt}
\alias{fun_extend_sig_emt}
\title{Calculate the EMT signature}
\usage{
fun_extend_sig_emt(input_mat, outfile)
}
\arguments{
\item{input_mat}{matrix}

\item{outfile}{output file}
}
\value{
data.frame with score
}
\description{
Calculate the EMT signature
}
