% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_sig.R
\name{fun_extend_sig_tip}
\alias{fun_extend_sig_tip}
\title{Signature deconvolution: TIP}
\usage{
fun_extend_sig_tip(input_mat, outfile)
}
\arguments{
\item{input_mat}{input matrix}

\item{outfile}{results saving files}
}
\value{
data.frame
}
\description{
Signature deconvolution: TIP
}
