% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_feature_selection_surv}
\alias{fun_feature_selection_surv}
\title{Survival feature selection}
\usage{
fun_feature_selection_surv(
  input_df,
  input_variables,
  outfile,
  pval_cutoff = 0.05,
  methods_remove = NULL,
  methods = c("lasso", "eNet75", "eNet50", "eNet25", "stepForward", "stepBackward",
    "stepBoth", "multicox", "unicox", "boost", "rf"),
  fun_train_cox_args = list(),
  fun_train_cox_step_args = list(),
  fun_surv_multicox_args = list(),
  fun_surv_unicox_args = list(),
  fun_train_cox_boost_args = list(),
  fun_train_cox_rf_args = list()
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{outfile}{the results saved on disk}

\item{pval_cutoff}{pval_cutoff for unicox and multicox}

\item{methods_remove}{not using the methods}

\item{methods}{"lasso","eNet75","eNet50","eNet25","stepForward","stepBackward","stepBoth","multicox","unicox","boost","rf"}

\item{fun_train_cox_args}{fun_train_cox_args}

\item{fun_train_cox_step_args}{fun_train_cox_step_args}

\item{fun_surv_multicox_args}{fun_surv_multicox_args}

\item{fun_surv_unicox_args}{fun_surv_unicox_args}

\item{fun_train_cox_boost_args}{fun_train_cox_boost_args}

\item{fun_train_cox_rf_args}{fun_train_cox_rf_args}
}
\value{
feature list
}
\description{
Survival feature selection
}
