% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_logit_multi}
\alias{fun_logit_multi}
\title{logistic-regression: multiple}
\usage{
fun_logit_multi(
  input_df,
  input_variables,
  input_y,
  return_pval = F,
  return_fit = F,
  glm_args = list()
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y}

\item{return_pval}{T}

\item{return_fit}{F}

\item{glm_args}{glm_args}
}
\value{
data.frame or list with fit and dataframe
}
\description{
logistic-regression: multiple
}
