% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_logit_uni}
\alias{fun_logit_uni}
\title{logistic regression}
\usage{
fun_logit_uni(
  input_df,
  input_variables,
  input_y,
  fun_logit_multi_args = list(return_pval = T)
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y}

\item{fun_logit_multi_args}{fun_logit_multi_args}
}
\value{
data.frame
}
\description{
logistic regression
}
