% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mat.R
\name{fun_mat_norm}
\alias{fun_mat_norm}
\title{matrix normalization}
\usage{
fun_mat_norm(
  input_mat,
  method = c("scale", "min-max", "towards"),
  axis = 1,
  towards_sample = NULL,
  fun_norm_scale_args = list(),
  fun_norm_minmax_args = list(),
  fun_norm_substract_args = list()
)
}
\arguments{
\item{input_mat}{input_matrix}

\item{method}{normalization methods, "scale","min-max","towards"}

\item{axis}{0 means row, 1 means column}

\item{towards_sample}{sampe_names used for towards}

\item{fun_norm_scale_args}{fun_norm_scale_args}

\item{fun_norm_minmax_args}{fun_norm_minmax_args}

\item{fun_norm_substract_args}{fun_norm_substract_args}
}
\value{
normalized matrix
}
\description{
matrix normalization
}
