% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mat.R
\name{fun_mat_qn}
\alias{fun_mat_qn}
\title{QN normalization}
\usage{
fun_mat_qn(input_mat, qn_args = list(keep.names = TRUE))
}
\arguments{
\item{input_mat}{input_mat}

\item{qn_args}{qn_args}
}
\value{
matrix
}
\description{
QN normalization
}
