% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mat.R
\name{fun_mat_rename2symbol}
\alias{fun_mat_rename2symbol}
\title{reanmes genes number into symbols}
\usage{
fun_mat_rename2symbol(input_mat, fun_gene2symbol_args = list())
}
\arguments{
\item{input_mat}{the input matrix}

\item{fun_gene2symbol_args}{the argument passing tp fun_gene2symbol from biodata}
}
\value{
matrix
}
\description{
reanmes genes number into symbols
}
