% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_merge.R
\name{fun_merge_df}
\alias{fun_merge_df}
\title{Merge the varaibles found in data.frame or file position, merge it into}
\usage{
fun_merge_df(
  input_df,
  input_df_merge,
  input_variables = NULL,
  input_regex = "xcell|cibersort|mcpcounter",
  key = "sample",
  df_merge_key = "cell_type"
)
}
\arguments{
\item{input_df}{input data.frame}

\item{input_df_merge}{the data.frame to merge}

\item{input_variables}{variables to merge}

\item{input_regex}{regex of variables to merge}

\item{key}{the input_df key, used to merge}

\item{df_merge_key}{the input_df_merge key, used to merge}
}
\value{
data.frame with merged varaibles
}
\description{
Merge the varaibles found in data.frame or file position, merge it into
}
