% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_merge.R
\name{fun_merge_maf}
\alias{fun_merge_maf}
\title{Merge the mutation status of gene into clinical data.frame}
\usage{
fun_merge_maf(input_df, input_maf, input_variables, key = "sample")
}
\arguments{
\item{input_df}{input_data.frame}

\item{input_maf}{the maf format data}

\item{input_variables}{genes to retrive the mutation data or list}

\item{key}{the key(sample,Tumor_Sample_Barcode)}
}
\value{
a data.frame with genes mutation status
}
\description{
Merge the mutation status of gene into clinical data.frame
}
