% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_merge.R
\name{fun_merge_matrix}
\alias{fun_merge_matrix}
\title{Merge varibles of matrix into dataframe}
\usage{
fun_merge_matrix(
  input_df,
  input_matrix,
  input_variables,
  output_variables_names = input_variables,
  key = "sample"
)
}
\arguments{
\item{input_df}{input_df}

\item{input_matrix}{input_matrix}

\item{input_variables}{variables to merge}

\item{key}{}
}
\value{
data.frame
}
\description{
Merge varibles of matrix into dataframe
}
