% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendelian_randomization.R
\name{fun_mr}
\alias{fun_mr}
\title{Mendelian randomization}
\usage{
fun_mr(
  exposure,
  outcome,
  res_rds,
  echo = T,
  mr_presso = T,
  fun_mr_available_outcomes_args = list(),
  fun_mr_extract_instrument_args = list(instrument_dir = "~/base/db/mr",
    extract_instruments_args = list(p1 = 1e-10)),
  extract_outcome_data_args = list()
)
}
\arguments{
\item{exposure}{exposure}

\item{outcome}{outcome}

\item{res_rds}{the results file}

\item{echo}{where print summary}

\item{mr_presso}{whether MR_presso}

\item{fun_mr_available_outcomes_args}{fun_mr_available_outcomes_args}

\item{fun_mr_extract_instrument_args}{fun_mr_extract_instrument_args}

\item{extract_outcome_data_args}{extract_outcome_data_args}
}
\value{
a list of MR results
}
\description{
Mendelian randomization
}
