% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendelian_randomization.R
\name{fun_mr_available_outcomes}
\alias{fun_mr_available_outcomes}
\title{retrive avaiable outcomes}
\usage{
fun_mr_available_outcomes(
  ao_file = "~/base/db/mr/available_outcomes.rds",
  available_outcomes_args = list()
)
}
\arguments{
\item{ao_file}{output file}

\item{available_outcomes_args}{args to available_outcomes}
}
\value{
df
}
\description{
retrive avaiable outcomes
}
