% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendelian_randomization.R
\name{fun_mr_extract_instrument}
\alias{fun_mr_extract_instrument}
\title{Get instruments SNPs}
\usage{
fun_mr_extract_instrument(
  exposure,
  instrument_dir = "~/base/db/mr",
  extract_instruments_args = list(p1 = 1e-10)
)
}
\arguments{
\item{exposure}{the exposure id}

\item{instrument_dir}{the output instrument-snps dir}

\item{extract_instruments_args}{the args}
}
\value{
df
}
\description{
Get instruments SNPs
}
