% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram.R
\name{fun_nomogram_logit}
\alias{fun_nomogram_logit}
\title{nomogram training for logistics regression}
\usage{
fun_nomogram_logit(
  input_df,
  input_variables,
  input_y,
  val_list = NULL,
  funlabel = "Probability",
  xfrac = 0.2,
  cex.axis = 0.5,
  cex.var = 0.8
)
}
\arguments{
\item{input_df}{the training data}

\item{input_variables}{the training varaibles}

\item{input_y}{the response data,one-hot encoding is required}

\item{val_list}{the predictive results,after training, predict the yhat for each validation datasets}

\item{xfrac}{nomogram plot function}

\item{cex.axis}{nomogram plot function}

\item{cex.var}{nomogtam plot funciton}
}
\value{
todo
}
\description{
nomogram training for logistics regression
}
