% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram.R
\name{fun_nomogram_predict}
\alias{fun_nomogram_predict}
\title{nomogram predict function}
\usage{
fun_nomogram_predict(input_df, input_fit, input_variables, input_y)
}
\arguments{
\item{input_df}{input data frame with each variables}

\item{input_fit}{the trained model}

\item{input_variables}{the variables used in fit}

\item{input_y}{the response data}
}
\value{
a predictive dataframe with yhat(which is the predictive results)
}
\description{
nomogram predict function
}
