% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{fun_norm_df}
\alias{fun_norm_df}
\title{data.frame normlization}
\usage{
fun_norm_df(
  input_df,
  input_variables,
  output_variable = NULL,
  method = c("scale", "min-max", "towards"),
  towards_sample = NULL,
  fun_norm_scale_args = list(),
  fun_norm_minmax_args = list(),
  fun_norm_substract_args = list()
)
}
\arguments{
\item{input_df}{data.frame obj}

\item{input_variables}{variables to normalized}

\item{output_variable}{output_variables for input_varaibles}

\item{method}{normalization methods, "scale","min-max","towards"}

\item{towards_sample}{towards_sample,currently not suppored}

\item{fun_norm_scale_args}{fun_norm_scale_args}

\item{fun_norm_minmax_args}{fun_norm_minmax_args}

\item{fun_norm_substract_args}{fun_norm_substract_args}
}
\value{
normlized dataframe
}
\description{
data.frame normlization
}
