% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{fun_norm_minmax}
\alias{fun_norm_minmax}
\title{min max normalization, change data int 0-1}
\usage{
fun_norm_minmax(input_numeric)
}
\arguments{
\item{input_numeric}{input_numeric data,DO NOT INPUT MATRIX}
}
\value{
normalized numeric
}
\description{
min max normalization, change data int 0-1
}
