% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{fun_norm_scale}
\alias{fun_norm_scale}
\title{rewrite scale function, remove all other attributes except normalized data}
\usage{
fun_norm_scale(input_numeric, scale_args = list())
}
\arguments{
\item{input_numeric}{input_data}

\item{scale_args}{args pass to scale function}
}
\value{
normalized data
}
\description{
rewrite scale function, remove all other attributes except normalized data
}
