% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{fun_norm_substract}
\alias{fun_norm_substract}
\title{normalized(been subtracted) input_mat1 by rowMeans of input_mat2}
\usage{
fun_norm_substract(input_mat1, input_mat2)
}
\arguments{
\item{input_mat1}{data used to normalize(kept)}

\item{input_mat2}{data used for normalizing(dropped)}
}
\value{
normalized input_mat1
}
\description{
normalized(been subtracted) input_mat1 by rowMeans of input_mat2
}
