% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_bar_pval}
\alias{fun_plot_bar_pval}
\title{barplot with pvalue}
\usage{
fun_plot_bar_pval(
  input_df,
  input_xs,
  input_ys,
  palette = "jco",
  bar_width = 0.3,
  tn = tn_bar(),
  chisq = F,
  fun_stat_fisher_by_column_args = list()
)
}
\arguments{
\item{input_df}{data.frame}

\item{input_xs}{x}

\item{input_ys}{y}

\item{bar_width}{bar withth}

\item{tn}{theme}
}
\value{
ggplot obj
}
\description{
barplot with pvalue
}
