% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_boxplot}
\alias{fun_plot_boxplot}
\title{boxplot}
\usage{
fun_plot_boxplot(
  input_df,
  input_x,
  input_y,
  palette = "jco",
  theme_now = tn(),
  remove_suffix = T,
  str_remove = "\\\\$|_",
  p_text_size = 3
)
}
\arguments{
\item{input_df}{data.frame}

\item{input_x}{group data}

\item{input_y}{ydata,regular expression}

\item{palette}{palette passing to fill_paltte in ggpubr}

\item{theme_now}{ggplot theme}

\item{str_remove}{ggtitle remove}

\item{p_text_size}{text size}
}
\value{
ggplot or list with ggplots
}
\description{
boxplot
}
