% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_butterfly}
\alias{fun_plot_butterfly}
\title{Butterfly plot for correlation anaysis}
\usage{
fun_plot_butterfly(
  input_df,
  input_key,
  input_variables_lower,
  input_variables_upper,
  return_plot_list = F,
  plot_text_size = 7,
  plot_x_angle = 75,
  plot_color_map_bre = c(-1, 0, 1),
  plot_color_map_col = c("#3A3A98", "white", "#832424"),
  tn_lgd = theme(legend.key.size = unit(5, "mm"), legend.position = "bottom", legend.text
    = element_text(size = 9), legend.title = element_text(size = 9), legend.spacing =
    unit(5, "mm"), legend.direction = "vertical"),
  plot_rect_col = "gray",
  plot_curvature = 0.1,
  tn = tn_empty(),
  lower_pos_shift = -2,
  upper_pos_shift = 2,
  larger_length = 3,
  smaller_length = 1.6,
  input_df_is_cor = F
)
}
\arguments{
\item{input_df}{data.frame or correlation results from \code{fun_stat_cor}}

\item{input_key}{key}

\item{input_variables_lower}{lower heatmap variables}

\item{input_variables_upper}{upper heatmap variables}

\item{return_plot_list}{whethre return plot list}

\item{plot_text_size}{plot_text_size}

\item{plot_x_angle}{plot_x_angle}

\item{plot_color_map_bre}{plot_color_map_bre}

\item{plot_color_map_col}{plot_color_map_col}

\item{tn_lgd}{tn_lgd}

\item{plot_rect_col}{plot_rect_col}

\item{plot_curvature}{plot_curvature}

\item{lower_pos_shift}{lower_pos_shift}

\item{upper_pos_shift}{upper_pos_shift}

\item{larger_length}{larger_length}

\item{smaller_length}{smaller_length}

\item{input_df_is_cor}{input_df_is_cor}
}
\value{
plot
}
\description{
Butterfly plot for correlation anaysis
}
