% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_density_box}
\alias{fun_plot_density_box}
\title{boxplot and density}
\usage{
fun_plot_density_box(
  input_df,
  x = "Group",
  y = "Risk_score",
  palette = "jco",
  heights = c(2, 1),
  tn_top = tn(),
  tn_bottom = tn()
)
}
\arguments{
\item{input_df}{data.frame}

\item{x}{x}

\item{y}{y}

\item{palette}{"jco"}

\item{heights}{ggarange heights}

\item{tn_top}{top plot theme}

\item{tn_bottom}{bottom plot theme}
}
\value{
plot
}
\description{
boxplot and density
}
