% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_forest}
\alias{fun_plot_forest}
\title{Forestplot}
\usage{
fun_plot_forest(
  input_df,
  x = "hr",
  y = "features",
  ci_low = "ci_low",
  ci_up = "ci_up",
  y_order = NULL,
  errorbar = T,
  segment = F,
  color_point = F,
  color_palette = "default",
  color_use = y,
  point_size = 1,
  point_shape = 15,
  vline_color = "gray"
)
}
\arguments{
\item{input_df}{input_df}

\item{x}{x}

\item{y}{y}

\item{ci_low}{lower confident interval}

\item{ci_up}{upper confident interval}

\item{y_order}{the y label order}

\item{errorbar}{T or F, whether using geom_errorbar function to plot the interval}

\item{segment}{T or F, whether using geom_segment function to plot the interval}

\item{color_point}{T or F, whether to color the point}

\item{color_palette}{the color palete passing to ggpubr::color_palette}

\item{color_use}{which features used to plot the color}

\item{point_size}{point size}

\item{point_shape}{point shape}

\item{vline_color}{vertical line(x=1)}
}
\value{
ggplot obj
}
\description{
Forestplot
}
