% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_forest2}
\alias{fun_plot_forest2}
\title{forest plot}
\usage{
fun_plot_forest2(
  input_df,
  x = "or",
  y = "Hugo_Symbol",
  ci_low = "ci.low",
  ci_up = "ci.up",
  feature1 = "High",
  feature2 = "Low",
  pval = "pval",
  order = NULL,
  desc = F,
  tn = tn_empty(),
  widths = c(2, 1, 3, 1.5)
)
}
\arguments{
\item{input_df}{input_df}

\item{x}{odds ratio or  hr}

\item{y}{genes}

\item{ci_low}{ci.low}

\item{ci_up}{ci.up}

\item{feature1}{features}

\item{feature2}{features}

\item{pval}{pval}

\item{order}{order}

\item{desc}{T or F}

\item{tn}{theme}

\item{widths}{the widths of ggarrange}
}
\value{
ggarrange obj
}
\description{
forest plot
}
