% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_heatmap}
\alias{fun_plot_heatmap}
\title{Heatmap plot using Heatmpa in ComplexHeatmap package}
\usage{
fun_plot_heatmap(
  input_matrix,
  top_df,
  left_df,
  name = "z-score",
  show_column_names = F,
  show_column_dend = F,
  cluster_column_slices = F,
  show_row_dend = F,
  row_title = " ",
  cluster_row_slices = F,
  row_name_cex = 0.5,
  row_name_face = "italic",
  row_name_pval_cutoff = 0.05,
  row_name_color_sig_up = "red",
  row_name_color_ns = "black",
  row_name_color_sig_down = "blue",
  top_split = 1,
  top_col = biodf::fun_utils_get_color(top_df),
  top_args = list(),
  left_split = 1,
  left_col = biodf::fun_utils_get_color(left_df),
  left_args = list(),
  hide_column_df_annotation = F,
  hide_row_df_annotation = F,
  ...
)
}
\arguments{
\item{input_matrix}{the plot matrix}

\item{top_df}{top_annotaiton data.frame,colnames should be in rownames of input_matrix}

\item{left_df}{left_annotation data.frame, rownames should be in rownames of input_matrix}

\item{name}{label the matrix heatmap legend title}

\item{show_column_names}{whether showing the column name}

\item{show_column_dend}{whether showing the  column dendro}

\item{cluster_column_slices}{whether cluster column slice}

\item{show_row_dend}{show_row_dend}

\item{row_title}{row_title, defualt is " ", showing nothing}

\item{cluster_row_slices}{cluster_row_slices}

\item{row_name_cex}{rownames font params, parssing to gp()}

\item{row_name_face}{rownames font params, parssing to gp()}

\item{row_name_pval_cutoff}{pvalue cutoff if the row_name was calculated}

\item{row_name_color_sig_up}{red}

\item{row_name_color_ns}{black}

\item{row_name_color_sig_down}{blue}

\item{top_split}{left split data, should be in colnames of left_df or number index}

\item{top_col}{list of top_annotation color map}

\item{top_args}{params list passing to HeatmapAnnotation}

\item{left_split}{left split data, should be in colnames of left_df or number index}

\item{left_col}{list of left_annotataion color map}

\item{left_args}{params list passing to RowAnnotation}

\item{hide_column_df_annotation}{F}

\item{hide_row_df_annotation}{F}

\item{...}{passing to ComplexHeatmap::Heatmap}
}
\value{
ComplexHeatmap plot
}
\description{
Heatmap plot using Heatmpa in ComplexHeatmap package
}
