% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_heatmap_tri}
\alias{fun_plot_heatmap_tri}
\title{Heatmap plot with fill and p-value text}
\usage{
fun_plot_heatmap_tri(
  input_cor_df,
  x = "x",
  y = "y",
  fill = "cor",
  text = "p_sig",
  Heatmap_args = list(cluster_rows = F, cluster_columns = F, name = "correlation",
    row_names_gp = gpar(fontsize = 8, fontface = "italic"), column_names_gp =
    gpar(fontsize = 8, fontface = "plain"), rect_gp = gpar(fill = "white", color =
    "gray"), col = circlize::colorRamp2(c(-0.5, 0, 0.5), colors =
    c(scales::muted("blue"), "white", scales::muted("red")))),
  Heatmap_args_additional = list(),
  hm_text_size = 7
)
}
\arguments{
\item{input_cor_df}{correlation reults}

\item{x}{x}

\item{y}{y}

\item{fill}{correlation}

\item{text}{pval}

\item{Heatmap_args}{other params passing to args}

\item{Heatmap_args_additional}{additional args into heatmap}
}
\value{
heatmap plot
}
\description{
Heatmap plot with fill and p-value text
}
