% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_roc.R
\name{fun_plot_roc}
\alias{fun_plot_roc}
\title{ROC plot}
\usage{
fun_plot_roc(
  input_roc,
  legend.position = c(0.7, 0.2),
  title = NULL,
  palette = c("black"),
  digits = 2
)
}
\arguments{
\item{input_roc}{input_roc list generated by fun_roc}

\item{legend.position}{legend.position}

\item{title}{title}

\item{palette}{palette}

\item{digits}{}
}
\description{
ROC plot
}
