% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_plot_surv_roc_time}
\alias{fun_plot_surv_roc_time}
\title{Plot time roc by the results of \code{fun_surv_roc}, extract the roc and auc}
\usage{
fun_plot_surv_roc_time(
  input_surv_roc_list,
  palette = "jco",
  title = "",
  legend_suffix = "-m survival ",
  legend_prefix = "AUC of ",
  legend.position = c(0.6, 0.2)
)
}
\arguments{
\item{input_surv_roc_list}{input_surv_roc_list}

\item{palette}{using get_paltte function in ggpubr, if its length equals to 1}

\item{title}{title}

\item{legend_suffix}{legend_suffix}

\item{legend_prefix}{legend_prefix}

\item{legend.position}{legend.position}
}
\value{
ggplot list
}
\description{
Plot time roc by the results of \code{fun_surv_roc}, extract the roc and auc
}
