% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot_volcano1}
\alias{fun_plot_volcano1}
\title{Volcanoplot}
\usage{
fun_plot_volcano1(
  input_df,
  pval = "adj.P.Val",
  pval_cutoff = 0.01,
  logFC = "logFC",
  logFC_cutoff = 1,
  palette = c("blue", "gray", "red"),
  tn = ggplot2::theme_minimal()
)
}
\arguments{
\item{input_df}{input_df}

\item{pval}{pval select}

\item{pval_cutoff}{pval cutoff}

\item{logFC}{logFC select}

\item{logFC_cutoff}{logFC selected}

\item{palette}{color}

\item{tn}{theme default}
}
\value{
ggplot obj
}
\description{
Volcanoplot
}
