% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read.R
\name{fun_read_cbio}
\alias{fun_read_cbio}
\title{read data into list}
\usage{
fun_read_cbio(input_clin = NULL, input_expr = NULL, input_maf = NULL)
}
\arguments{
\item{input_clin}{clinical file postion}

\item{input_expr}{expression file postion}

\item{input_maf}{maf file position}
}
\value{
a list contain data
}
\description{
read data into list
}
