% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read.R
\name{fun_read_df}
\alias{fun_read_df}
\title{read file into dataframe}
\usage{
fun_read_df(input_file, index_col = "sample", ...)
}
\arguments{
\item{input_file}{the file postion of large text}

\item{index_col}{which column is index}

\item{...}{params pass to fread}
}
\value{
data.frame objc
}
\description{
read file into dataframe
}
