% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read.R
\name{fun_read_df2maf}
\alias{fun_read_df2maf}
\title{Change the data.frame format into maf format}
\usage{
fun_read_df2maf(
  input_df,
  ...,
  Hugo_Symbol = "Hugo_Symbol",
  Chromosome = "Chromosome",
  Start_Position = "Start_Position",
  End_Position = "End_Position",
  Reference_Allele = "Reference_Allele",
  Tumor_Seq_Allele2 = "Tumor_Seq_Allele2",
  Variant_Classification = "Variant_Type",
  Variant_Type = "Variant_Type",
  Tumor_Sample_Barcode = "Tumor_Sample_Barcode"
)
}
\arguments{
\item{input_df}{data.frame or file position of maf file}

\item{...}{other fields for maf column}

\item{Hugo_Symbol}{Hugo_Symbol}

\item{Chromosome}{Chromosome}

\item{Start_Position}{Start_Position}

\item{End_Position}{End_Position}

\item{Reference_Allele}{Reference_Allele}

\item{Tumor_Seq_Allele2}{Tumor_Seq_Allele2}

\item{Variant_Classification}{Variant_Classification}

\item{Variant_Type}{Variant_Type}

\item{Tumor_Sample_Barcode}{Tumor_Sample_Barcode}
}
\value{
a maf-like format
}
\description{
Change the data.frame format into maf format
}
