% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read.R
\name{fun_read_meth_idat}
\alias{fun_read_meth_idat}
\title{read methylation idat from dir, return the beta metarix}
\usage{
fun_read_meth_idat(
  input_dir,
  outfile,
  sample_anno_file = "sample_idat_file.csv",
  input_df = NULL,
  input_sample = "sample",
  input_filename = "filename",
  input_sample_regex = "GSM\\\\d+",
  idat_suffix_regex =
    "_Grn\\\\.idat$|_Grn\\\\.idat\\\\.gz$|_Red\\\\.idat$|_Red\\\\.idat\\\\.gz$"
)
}
\arguments{
\item{input_dir}{input_dir contains the idat files(Grn,Red)}

\item{outfile}{the beta matrix file position}

\item{sample_anno_file}{if supplied write annotation file}

\item{input_df}{data.frame contains sample name, and idat file name}

\item{input_sample}{sample_colmn}

\item{input_filename}{idata filename}

\item{input_sample_regex}{extract sample form idat filename, if input_df is NULL}

\item{idat_suffix_regex}{extract idat form input_dir, if input_df is NULL}
}
\value{
data.frame
}
\description{
read methylation idat from dir, return the beta metarix
}
