% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read.R
\name{fun_read_tisch}
\alias{fun_read_tisch}
\title{read single cell data downloaded from}
\usage{
fun_read_tisch(input_dir, output_file)
}
\arguments{
\item{input_dir}{input directory}

\item{output_file}{the saved filed}
}
\value{
list of single cell
}
\description{
read single cell data downloaded from
}
