% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{fun_relabel_vec_count}
\alias{fun_relabel_vec_count}
\title{Rename the vector the the count of each unique value, output is the same lenght as input}
\usage{
fun_relabel_vec_count(input_vec, count_prefix = "(", count_suffix = ")")
}
\arguments{
\item{input_vec}{the input vector of labels}

\item{count_prefix}{prefix}

\item{count_suffix}{suffix}
}
\value{
input vector
}
\description{
Rename the vector the the count of each unique value, output is the same lenght as input
}
