% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run.R
\name{fun_run_gistic2}
\alias{fun_run_gistic2}
\title{Run gistic2}
\usage{
fun_run_gistic2(
  input_df,
  input_cnv_seg_file,
  output_dir,
  input_group_name = "Group",
  input_sample_name = "sample",
  loading = T,
  gistic2_command = "/home/r/miniconda3/envs/gistic2/bin/gistic2",
  gistic2_refgene =
    "/home/r/miniconda3/envs/gistic2/share/gistic2-2.0.23-0/refgenefiles/hg19.mat",
  LD_LIBRARY_PATH =
    "/home/r/miniconda3/envs/gistic2/share/mcr-8.3-0/v83/runtime/glnxa64:/home/r/miniconda3/envs/gistic2/share/mcr-8.3-0/v83/bin/glnxa64:/home/r/miniconda3/envs/gistic2/share/mcr-8.3-0/v83/sys/os/glnxa64:"
)
}
\arguments{
\item{input_df}{meta-data, used to extract data}

\item{input_cnv_seg_file}{cnv segment file}

\item{output_dir}{The results folder}

\item{input_group_name}{Group}

\item{input_sample_name}{sample}

\item{loading}{whether reading the results using maftools::readGistic}

\item{gistic2_command}{gistic2 commaon}

\item{gistic2_refgene}{gistic2 refegene}

\item{LD_LIBRARY_PATH}{gistic2 LD_LIBRARY_PATH}
}
\value{
gistic2
}
\description{
Run gistic2
}
