% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_cellchat}
\alias{fun_sc_cellchat}
\title{cell_chat analysis}
\usage{
fun_sc_cellchat(
  input_sc,
  cellchat_rds,
  group.by = "cell_type",
  cell_search = "Secreted Signaling",
  workers = 4
)
}
\arguments{
\item{input_sc}{Seurat obj}

\item{cellchat_rds}{the output file used to save results}

\item{group.by}{cellchat idents}

\item{cell_search}{cell db use}

\item{workers}{parrallel}
}
\value{
cellchat obj
}
\description{
cell_chat analysis
}
