% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_cellchat_netVisual_circle}
\alias{fun_sc_cellchat_netVisual_circle}
\title{plot cell chat}
\usage{
fun_sc_cellchat_netVisual_circle(
  input_cellchat,
  input_cell = NULL,
  use_used = "count",
  netVisual_circle_args = list(weight.scale = T)
)
}
\arguments{
\item{input_cellchat}{cell chat}

\item{input_cell}{which cell to plot, NULL or cell types}

\item{use_used}{count or weight}

\item{netVisual_circle_args}{}
}
\description{
plot cell chat
}
