% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_cellchat_plot}
\alias{fun_sc_cellchat_plot}
\title{plot cell chat}
\usage{
fun_sc_cellchat_plot(input_cellchat, input_cell = NULL)
}
\arguments{
\item{input_cellchat}{cell chat}

\item{input_cell}{which cell to plot, NULL or cell types}
}
\description{
plot cell chat
}
