% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_plot_scissor}
\alias{fun_sc_plot_scissor}
\title{Scissor plot}
\usage{
fun_sc_plot_scissor(
  input_sc,
  pt.size = 0.8,
  group.by = "scissor",
  cols = c("grey", "indianred1", "royalblue")
)
}
\arguments{
\item{input_sc}{seurat plot}

\item{pt.size}{pt.size}

\item{group.by}{group.by}

\item{cols}{colors}
}
\value{
ggplot obj
}
\description{
Scissor plot
}
