% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_read_mtx}
\alias{fun_sc_read_mtx}
\title{fun_sc_read_mtx}
\usage{
fun_sc_read_mtx(
  input_dir,
  outfile,
  input_anno = NULL,
  mtx_file_search = "barcodes.tsv.gz"
)
}
\arguments{
\item{input_dir}{directory contains the mtx file}

\item{outfile}{the outfile}

\item{input_anno}{data.frame, first column is sample name, second column is the mtx file position}

\item{mtx_file_search}{key word to find mtx dir position}
}
\value{
todo
}
\description{
fun_sc_read_mtx
}
