% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_seurat_preprocessing}
\alias{fun_sc_seurat_preprocessing}
\title{Single cell preprocessing wrapper, params set to NULL to skip}
\usage{
fun_sc_seurat_preprocessing(
  input_sc,
  filter_feature_low = NULL,
  filter_feature_high = NULL,
  filter_mt = NULL,
  scale_factor = NULL,
  variables_method = "vst",
  variables_feature = 2000,
  scale_data = T,
  nPCA = 50,
  clustering_dims = 1:15,
  clustering_resolution = 0.5,
  umap_dims = 1:15
)
}
\arguments{
\item{input_sc}{input_seurat project or input matirx}

\item{filter_feature_low}{the minimum feature}

\item{filter_feature_high}{the maximum feature}

\item{filter_mt}{the mitocondrial gens cutoff}

\item{scale_factor}{the scale factor}

\item{variables_method}{variables_method}

\item{variables_feature}{variables_feature}

\item{scale_data}{scale_data}

\item{nPCA}{nPCA}

\item{clustering_dims}{clustering_dims}

\item{clustering_resolution}{clustering_resolution}

\item{umap_dims}{umap_dims}
}
\description{
Single cell preprocessing wrapper, params set to NULL to skip
}
