% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_seurat_progeny}
\alias{fun_sc_seurat_progeny}
\title{Calculate the progeny score for Seurat obj using decouplerR}
\usage{
fun_sc_seurat_progeny(
  input_seurat,
  out_file,
  fun_sc = function(x) x@assays$RNA@data,
  top = 100
)
}
\arguments{
\item{input_seurat}{seurat obj}

\item{out_file}{the calculated score output}

\item{fun_sc}{which expression data in Seurat to use, the default is RNA data}

\item{top}{how many progeny genes for each pathway to use, default is 100}
}
\description{
Calculate the progeny score for Seurat obj using decouplerR
}
