% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_seurat_scissor}
\alias{fun_sc_seurat_scissor}
\title{Scissor analysis}
\usage{
fun_sc_seurat_scissor(...)
}
\arguments{
\item{...}{params passing to Scissor::Scissor}
}
\value{
seurat obj with scissor feature
}
\description{
Scissor analysis
}
