% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_seurat_trajectory}
\alias{fun_sc_seurat_trajectory}
\title{Plot trajectory using monocle2}
\usage{
fun_sc_seurat_trajectory(
  input_sc,
  input_genes,
  input_stat = "cell_major",
  input_root = NULL
)
}
\arguments{
\item{input_sc}{input_seurat}

\item{input_genes}{the genes}

\item{input_stat}{the cell status}

\item{input_root}{the root cell}
}
\value{
monocle single cell data obj
}
\description{
Plot trajectory using monocle2
}
