% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sc.R
\name{fun_sc_seurat_ucell}
\alias{fun_sc_seurat_ucell}
\title{Seurat obj signature deconvolution usingUCell}
\usage{
fun_sc_seurat_ucell(input_seurat, input_list, ucell_params = list(name = ""))
}
\arguments{
\item{input_seurat}{the Seurat object}

\item{input_list}{signature list, must has name}

\item{ucell_params}{parmaters passing to AddModuleScore_UCell}
}
\value{
Seurat obj with signature deconvoluted
}
\description{
Seurat obj signature deconvolution usingUCell
}
