% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{fun_sig_deconv}
\alias{fun_sig_deconv}
\title{Signature deconvolution,call function from \code{decoupleR}}
\usage{
fun_sig_deconv(
  input_mat,
  input_sig,
  outfile,
  .source = "term",
  .target = "gene",
  input_args = list(),
  to_matrix = T,
  statistic_used = NULL,
  run_method = "run_gsva"
)
}
\arguments{
\item{input_mat}{the data matrix}

\item{input_sig}{signature, data.frame}

\item{outfile}{outputfile}

\item{.source}{terms/source, stand for gene set names}

\item{.target}{genes column}

\item{input_args}{other arguments passing to run_method}

\item{to_matrix}{whether convert the results into matrix}

\item{statistic_used}{which stats used for matrix, only selected when using to_matrix}

\item{run_method}{method exported by \code{decoupleR}}
}
\value{
a tibble, or matrix
}
\description{
Signature deconvolution,call function from \code{decoupleR}
}
