% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{fun_sig_deonvolute_immune}
\alias{fun_sig_deonvolute_immune}
\title{Immune deonvolution}
\usage{
fun_sig_deonvolute_immune(
  input_expr,
  outfile,
  indications,
  tumor = T,
  arrays = F
)
}
\arguments{
\item{input_expr}{expression matrix, should not be log2 transformed}

\item{outfile}{outfile postion}

\item{indications}{cancer type}

\item{tumor}{T or F}

\item{arrays}{T or F}
}
\value{
data.table
}
\description{
Immune deonvolution
}
