% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{fun_sig_ssGSGA}
\alias{fun_sig_ssGSGA}
\title{ssGSEA analysis for input_mat using input_genes_list, the results would be saved as outfile}
\usage{
fun_sig_ssGSGA(input_mat, input_genes_list, outfile)
}
\arguments{
\item{input_mat}{input_mat}

\item{input_genes_list}{input_genes_list}

\item{outfile}{}
}
\value{
data.frame
}
\description{
ssGSEA analysis for input_mat using input_genes_list, the results would be saved as outfile
}
