% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{fun_sig_weighted_sum}
\alias{fun_sig_weighted_sum}
\title{Calculated the weight sum of the given input coefficients}
\usage{
fun_sig_weighted_sum(
  input_df,
  input_coef,
  coef_column = 1,
  output_name = "Risk_score"
)
}
\arguments{
\item{input_df}{input_df}

\item{input_coef}{input_coef}

\item{coef_column}{which coef to use}

\item{output_name}{the output name}
}
\value{
data.frame with calculated value
}
\description{
Calculated the weight sum of the given input coefficients
}
