% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_compare.R
\name{fun_stat_compare_means}
\alias{fun_stat_compare_means}
\title{Perform statistical test for variables and each group}
\usage{
fun_stat_compare_means(
  input_df,
  input_groups = "Group",
  input_variables = NULL,
  fun = "kruskal.test",
  fun_sum = "median"
)
}
\arguments{
\item{input_df}{input data.frame}

\item{input_groups}{input groups}

\item{input_variables}{input variables}

\item{fun}{test function, like \code{t.test}, \code{wilcox.test},\code{kruskal.test}}

\item{fun_sum}{how to summarise data for each group}
}
\value{
data.frame with pvalue, and summarise each group parms
}
\description{
Perform statistical test for variables and each group
}
