% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_compare.R
\name{fun_stat_diagnostic_test}
\alias{fun_stat_diagnostic_test}
\title{diagnostic test}
\usage{
fun_stat_diagnostic_test(
  input_df,
  input_variables,
  input_golden_standard,
  input_pos = "Positive",
  input_neg = "Negative",
  gold_pos = "Positive",
  gold_neg = "Negative"
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_golden_standard}{input_golden_standard}

\item{input_pos}{input_pos}

\item{input_neg}{input_neg}

\item{gold_pos}{gold_pos}

\item{gold_neg}{gold_neg}
}
\value{
df
}
\description{
diagnostic test
}
