% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fisher.R
\name{fun_stat_fisher_by_row}
\alias{fun_stat_fisher_by_row}
\title{Fisher test by row}
\usage{
fun_stat_fisher_by_row(input_df, input_abcd = c(2, 3, 4, 5))
}
\arguments{
\item{input_df}{input_data}

\item{input_abcd}{the abcd column}
}
\value{
data.frame which appedn the or, ci_low, ci_up, pval column
}
\description{
Fisher test by row
}
