% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fisher.R
\name{fun_stat_fisher_mat}
\alias{fun_stat_fisher_mat}
\title{Perform the fisher test, return the data.frame}
\usage{
fun_stat_fisher_mat(input_mat, input_name)
}
\arguments{
\item{input_mat}{the input matrix}

\item{input_name}{the name of the names(features and rownames of the output)}
}
\value{
data.frame summary of fisher.test results
}
\description{
Perform the fisher test, return the data.frame
}
