% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_concordance}
\alias{fun_surv_concordance}
\title{Calculate concordance step}
\usage{
fun_surv_concordance(input_df, input_variables, input_ref = as.integer(1))
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_ref}{set 1}
}
\value{
a data.frame with value
}
\description{
Calculate concordance step
}
