% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_cutoff}
\alias{fun_surv_cutoff}
\title{Group numeric variables into high-low groups, based on the cutoff(Lower side included)}
\usage{
fun_surv_cutoff(
  input_df,
  input_variables = "Risk_score",
  output_variables_names = "Group",
  input_pct = 0.5,
  label = c("Low", "High"),
  factor_level = T
)
}
\arguments{
\item{input_df}{Input is data.frame-like object}

\item{input_variables}{Numeric variables}

\item{output_variables_names}{output names}

\item{input_pct}{range(0,1), If length(input_pct) == 1, using it as quantitle as cutoff;if length(input_pct) ==2, group higher than max percentage -> High, lower than the min percentage -> Low; If length(input_pct) >3, group data with minumu pvalue(survival, should have \code{time},\code{status} columns)}

\item{label}{output labels, default: c("Low","High")}

\item{factor_level}{whether refactor group data, default is True}
}
\value{
A data.frame whose \code{input_variables} are grouped
}
\description{
Group numeric variables into high-low groups, based on the cutoff(Lower side included)
}
\examples{
data(tcga_clin)
clin_group = fun_surv_cutoff(
  tcga_clin,
  input_variables="age",
  output_variables_names = "Group",
  input_pct = 0.5)
}
