% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_multicox}
\alias{fun_surv_multicox}
\title{Multicox survival model, make sure you have time and status column, and label properly}
\usage{
fun_surv_multicox(input_df, input_variables = "Risk_score")
}
\arguments{
\item{input_df}{input data.frame, with variables and time, status columns}

\item{input_variables}{input variables}
}
\value{
data.frame-like objects
}
\description{
Multicox survival model, make sure you have time and status column, and label properly
}
\examples{
data(tcga_clin)
fun_surv_multicox(head(tcga_clin,100),c("age","Age_group"))
}
