% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_nomogram}
\alias{fun_surv_nomogram}
\title{Nomogram}
\usage{
fun_surv_nomogram(
  input_df,
  input_variables,
  times = c(12, 24, 36, 48, 60),
  B = 10,
  m = 50,
  lim = c(0, 1),
  time_suffix = "-month",
  outfile = "./plot/nomogram"
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{times}{times}

\item{B}{boostrap}

\item{m}{m}

\item{lim}{lim}

\item{time_suffix}{months}

\item{outfile}{pdf save postion}
}
\value{
todo
}
\description{
Nomogram
}
