% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_parsing}
\alias{fun_surv_parsing}
\title{Parse survival time}
\usage{
fun_surv_parsing(
  input_df,
  input_surv_index = c("OS.time", "OS", "1"),
  input_na_value = NULL
)
}
\arguments{
\item{input_df}{input_df}

\item{input_surv_index}{three-length vector}

\item{input_na_value}{label some labels with NA value}
}
\value{
data.frame like obj with \code{time} and \code{status} column
}
\description{
Parse survival time
}
