% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_parsing_shortcut}
\alias{fun_surv_parsing_shortcut}
\title{clinical data survival parsing quick entry}
\usage{
fun_surv_parsing_shortcut(
  input_df,
  shortcut = "pfs",
  fun_surv_parsing_arge = list()
)
}
\arguments{
\item{input_df}{clinical data}

\item{shortcut}{which data to parse}

\item{fun_surv_parsing_arge}{args passing to fun_surv_parsing}
}
\value{
data.frame: clinical data
}
\description{
clinical data survival parsing quick entry
}
