% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_roc}
\alias{fun_surv_roc}
\title{roc analysis for survival data}
\usage{
fun_surv_roc(
  input_df,
  input_variables = "Risk_score",
  input_times = c(1, 2, 3, 4, 5) * 12,
  input_ds_name = "Not supplied",
  merge_roc_auc = T
)
}
\arguments{
\item{input_df}{input.frame data}

\item{input_variables}{input variables}

\item{input_times}{times for survival time analysis}

\item{input_ds_name}{ds_names}

\item{merge_roc_auc}{merge input variables roc/auc in one data.frame respectively}
}
\value{
if \code{merge_roc_auc} is \code{TRUE},1)list with two data.frame(roc and auc) 2)Otherwise, lists with same length to input_variables
}
\description{
roc analysis for survival data
}
\examples{
# merge the output into two dataframe
res_list =
 fun_surv_roc(
 tcga_clin,
 input_variables = "age",
 merge_roc_auc=T )

}
