% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_surv.R
\name{fun_surv_unicox}
\alias{fun_surv_unicox}
\title{Unicox for survival model, make sure you have time and status column, and label properly}
\usage{
fun_surv_unicox(input_df, input_variables = "Risk_score", input_pct = NULL)
}
\arguments{
\item{input_df}{Input data.frame, with variables and time, status columns}

\item{input_variables}{Input variables}
}
\value{
data.frame-like objects
}
\description{
Unicox for survival model, make sure you have time and status column, and label properly
}
\examples{
data(tcga_clin)
fun_surv_unicox(head(tcga_clin,100),"Age_group")

}
