% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_train_cox}
\alias{fun_train_cox}
\title{Cox training}
\usage{
fun_train_cox(
  input_df,
  input_variables,
  input_y = c("time", "status"),
  alpha = 1,
  return_fit = F,
  cv.glmnet_args = list(type.measure = "C"),
  seed = 1,
  lambda = "lambda.min"
)
}
\arguments{
\item{input_df}{input_ff}

\item{input_variables}{variables for cox regression}

\item{input_y}{time and status column}

\item{alpha}{1->lasso, 0 -> ridge elastic (0,1)}

\item{return_fit}{whether return fit}

\item{cv.glmnet_args}{glmnet args}

\item{seed}{reproducible seed}

\item{lambda}{coefficients lambda}
}
\value{
coefficients coef or list with fit
}
\description{
Cox training
}
