% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_train_cox_boost}
\alias{fun_train_cox_boost}
\title{CoxBoost}
\usage{
fun_train_cox_boost(
  input_df,
  input_variables,
  input_y = c("time", "status"),
  return_fit = F,
  seed = 1
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{\code{time} and \code{status}}

\item{return_fit}{whether return fit}

\item{seed}{reproducible seed}
}
\value{
coef aor list with fit and coef
}
\description{
CoxBoost
}
