% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_train_cox_rf}
\alias{fun_train_cox_rf}
\title{Survival random forest}
\usage{
fun_train_cox_rf(
  input_df,
  input_variables,
  input_y = c("time", "status"),
  nsplit = 50,
  nodesize = 5,
  conservative = "low",
  return_fit = F,
  seed = 1
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{\code{time},\code{status}}

\item{nsplit}{pass to \code{randomForestSRC::rfsrc}}

\item{nodesize}{pass to \code{randomForestSRC::rfsrc}}

\item{conservative}{pass to \code{randomForestSRC::rfsrc}}

\item{return_fit}{whether return fit}

\item{seed}{reproducible seed}
}
\value{
filtered variables or list with fit and filtered variables
}
\description{
Survival random forest
}
