% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_train_cox_step}
\alias{fun_train_cox_step}
\title{Step cox analysis}
\usage{
fun_train_cox_step(
  input_df,
  input_variables,
  direction = "forward",
  input_y = c("time", "status"),
  return_fit = F
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{direction}{\code{both}, \code{forward} or \code{backward}}

\item{input_y}{\code{time} and \code{status} column, please set to \code{time} and \code{status} all the time}

\item{return_fit}{whether return fit}
}
\value{
a list with fit and coef, or just coef only
}
\description{
Step cox analysis
}
