% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_logit}
\alias{fun_train_logit}
\title{Training data using lasso-logistic}
\usage{
fun_train_logit(
  input_df,
  input_variables,
  input_y,
  lambda = "lambda.min",
  seed = 1,
  return_fit = F
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{(0,1)}

\item{lambda}{lambda}

\item{seed}{seed}

\item{return_fit}{whether return fit}
}
\value{
model_coef, or list with fit
}
\description{
Training data using lasso-logistic
}
