% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_logit_gbm}
\alias{fun_train_logit_gbm}
\title{logistic-regression: gbm}
\usage{
fun_train_logit_gbm(
  input_df,
  input_variables,
  input_y,
  seed = 1,
  return_fit = F,
  distribution = "bernoulli",
  n.trees = 100,
  interaction.depth = 1,
  cv.folds = 10,
  gbm_args = list()
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y}

\item{seed}{seed}

\item{return_fit}{return_fit}

\item{distribution}{the distribution of y}

\item{n.trees}{how many trees, default 100}

\item{interaction.depth}{interaction.depth}

\item{cv.folds}{cv: defulat 10}

\item{gbm_args}{#' @param cv.folds}
}
\value{
vector or gbm object
}
\description{
logistic-regression: gbm
}
