% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_logit_rf}
\alias{fun_train_logit_rf}
\title{train data with randomForest}
\usage{
fun_train_logit_rf(
  input_df,
  input_variables,
  input_y,
  seed = 1,
  return_fit = F,
  top_feature = 10,
  importance_method = "MeanDecreaseGini",
  randomForest_args = list(importance = TRUE)
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y, numeric for regression, factors for classification}

\item{seed}{seed}

\item{return_fit}{return_fit}

\item{top_feature}{The most important feature, default top 10}

\item{importance_method}{important methods, MeanDecreaseGini or MeanDecreaseAccuracy}

\item{randomForest_args}{other prams passing randomForst function}
}
\value{
train results of list or feature names
}
\description{
train data with randomForest
}
