% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_logit_step}
\alias{fun_train_logit_step}
\title{logistic-regression: step}
\usage{
fun_train_logit_step(
  input_df,
  input_variables,
  input_y,
  direction,
  return_fit = F
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y}

\item{direction}{direction}

\item{return_fit}{return_fit}
}
\value{
data.frame or list with fit and data.frame and
}
\description{
logistic-regression: step
}
