% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_logit_svm}
\alias{fun_train_logit_svm}
\title{svm-regression}
\usage{
fun_train_logit_svm(
  input_df,
  input_variables,
  input_y,
  seed = 1,
  cost = 1,
  scale = F,
  kernel = "linear",
  return_fit = F,
  svm_args = list()
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y}

\item{seed}{seed}

\item{cost}{cost: default 1}

\item{scale}{scale: F}

\item{kernel}{default: linear, consider radial}

\item{return_fit}{T}

\item{svm_args}{smv_args}
}
\value{
data.frame or list with fit and data.frame
}
\description{
svm-regression
}
