% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_train_cox.R
\name{fun_train_multiple_cox}
\alias{fun_train_multiple_cox}
\title{Train step, use with fun_feature_selection_surv and fun_design_train_combn}
\usage{
fun_train_multiple_cox(
  input_df,
  design_df,
  input_features_list,
  outfile,
  feature_selection_method = "m1",
  train_method = "m2",
  fun_train_cox_args = list(),
  fun_train_cox_step_args = list(),
  fun_surv_multicox_args = list(),
  fun_train_cox_boost_args = list()
)
}
\arguments{
\item{input_df}{input_df}

\item{design_df}{two column df, feature selection methods and training methods}

\item{input_features_list}{the feature selected methods list}

\item{outfile}{the results saving outfile}

\item{feature_selection_method}{the colname of feature selection in  design_df}

\item{train_method}{the colname of feature selection in  train method}

\item{fun_train_cox_args}{fun_train_cox}

\item{fun_train_cox_step_args}{fun_train_cox_args}

\item{fun_surv_multicox_args}{fun_surv_multicox}

\item{fun_train_cox_boost_args}{fun_train_cox_boost}
}
\value{
list of model coef
}
\description{
Train step, use with fun_feature_selection_surv and fun_design_train_combn
}
