% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_multiple_logit}
\alias{fun_train_multiple_logit}
\title{logit-train, with multiple methods}
\usage{
fun_train_multiple_logit(
  input_df,
  input_y,
  design_df,
  input_features_list,
  outfile,
  feature_selection_method = "m1",
  train_method = "m2",
  fun_train_logit_args = list(),
  fun_train_logit_step_args = list(),
  fun_logit_multi_args = list(),
  fun_train_logit_svm_args = list()
)
}
\arguments{
\item{input_df}{todo}

\item{input_y}{todo}

\item{design_df}{todo}

\item{input_features_list}{todo}

\item{outfile}{todo}

\item{feature_selection_method}{todo}

\item{train_method}{todo}

\item{fun_train_logit_args}{todo}

\item{fun_train_logit_step_args}{todo}

\item{fun_logit_multi_args}{todo}

\item{fun_train_logit_svm_args}{todo}
}
\value{
list of models
}
\description{
logit-train, with multiple methods
}
