% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_broadcast}
\alias{fun_utils_broadcast}
\title{This is broadcast function,change the shape of input_targets so that could fit into input_variables}
\usage{
fun_utils_broadcast(input_variables, input_targets)
}
\arguments{
\item{input_variables}{The input variables}

\item{input_targets}{The output with same shape as input variables}
}
\value{
return a vector of the input_variables with same shape
}
\description{
This is broadcast function,change the shape of input_targets so that could fit into input_variables
}
