% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_capital}
\alias{fun_utils_capital}
\title{Capital the each character}
\usage{
fun_utils_capital(input_strings, capital_position = 1, capital_length = 1)
}
\arguments{
\item{input_strings}{the strings}

\item{capital_position}{which character to captitolize}

\item{capital_length}{the length}
}
\value{
vector
}
\description{
Capital the each character
}
