% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_df2oncoprint}
\alias{fun_utils_df2oncoprint}
\title{Oncoprint data preparation}
\usage{
fun_utils_df2oncoprint(
  input_df,
  input_variables,
  input_variables_column = "Hugo_Symbol",
  input_mut = "mutation_type",
  input_sample = "sample",
  input_sample_append = NULL,
  fun_aggre = function(x) unique(x) \%>\% na.omit() \%>\% paste0(collapse = ";"),
  input_col = NULL,
  bg_col = "gray",
  cell_height = 1,
  cell_width = 0.9,
  height_decrease = 0.05
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{genes to visulize}

\item{input_variables_column}{gene column, like Hugo_Symbol in maf format}

\item{input_mut}{the mutation type column}

\item{input_sample}{the sample column, like Tumor_Sample_Barcode in maf format}

\item{input_sample_append}{input_sample_append}

\item{fun_aggre}{fun_aggre}

\item{input_col}{color for default mutation type}

\item{bg_col}{bg_col}

\item{cell_width}{cell_width}

\item{height_decrease}{height_decrease}
}
\value{
todo
}
\description{
Oncoprint data preparation
}
