% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_eval}
\alias{fun_utils_eval}
\title{eval the results}
\usage{
fun_utils_eval(input_texts)
}
\arguments{
\item{input_texts}{the text inputs}
}
\value{
the val res
}
\description{
eval the results
}
\examples{
a = c("1+2", "1/2", "3")
fun_utils_eval(a)
}
