% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r_env.R
\name{fun_utils_export_enviroment}
\alias{fun_utils_export_enviroment}
\title{export current enviroment}
\usage{
fun_utils_export_enviroment(outfile, lib.loc = NULL)
}
\arguments{
\item{outfile}{the csv files for storing packages}

\item{lib.loc}{}
}
\value{
data.frame
}
\description{
export current enviroment
}
