% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_get_color}
\alias{fun_utils_get_color}
\title{Get colors}
\usage{
fun_utils_get_color(
  input_df,
  input_variables = NULL,
  input_colors = NULL,
  color_shift = 4
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_colors}{input_colors}

\item{color_shift}{color_shift}
}
\value{
a list with color map
}
\description{
Get colors
}
