% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_insert_newline_for_long_stings}
\alias{fun_utils_insert_newline_for_long_stings}
\title{If the strins is to loner, say, 50 character, we would like to insert a new line into the strings, so that it could be printed beatifully}
\usage{
fun_utils_insert_newline_for_long_stings(
  input_strings,
  string_check = 50,
  sep = "_"
)
}
\arguments{
\item{input_strings}{the stings list}

\item{string_check}{integer indicates the long strings}

\item{sep}{how the stings separate}
}
\value{
a list of stings with newline inserted
}
\description{
If the strins is to loner, say, 50 character, we would like to insert a new line into the strings, so that it could be printed beatifully
}
