% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r_env.R
\name{fun_utils_install_enviroment}
\alias{fun_utils_install_enviroment}
\title{install packages from a file exported by fun_utils_export_enviroment}
\usage{
fun_utils_install_enviroment(
  pkg_enviroment_file,
  failed_files,
  cran_version = F,
  cran_repos_url = "https://mirrors.bfsu.edu.cn/CRAN/",
  proxy = "socks5://127.0.0.1:1080"
)
}
\arguments{
\item{pkg_enviroment_file}{the outfile of fun_utils_export_enviroment}

\item{failed_files}{packges not saved}

\item{cran_version}{T or F, whether install version}

\item{cran_repos_url}{repos}

\item{proxy}{proxy used for biocmanager}
}
\description{
install packages from a file exported by fun_utils_export_enviroment
}
