% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_p2star}
\alias{fun_utils_p2star}
\title{Conver tpval into star symbol}
\usage{
fun_utils_p2star(input_pval, ns = "")
}
\arguments{
\item{input_pval}{the input value}

\item{ns}{non-significant labels}
}
\value{
a vector of strings
}
\description{
Conver tpval into star symbol
}
