% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_split}
\alias{fun_utils_split}
\title{split data into train-test, or differents parts}
\usage{
fun_utils_split(
  input_data,
  input_pct = c(0.7, 0.3),
  key = "sample",
  using = "expr",
  seed = 1,
  return_sample = F
)
}
\arguments{
\item{input_data}{input data, could be df,matrix,ds_list}

\item{input_pct}{train-test percentage, 0.7,0.3}

\item{key}{key_column}

\item{using}{ds_list, using matrix}

\item{seed}{randdome seed}

\item{return_sample}{return samples list or same with input_data}
}
\value{
split obj
}
\description{
split data into train-test, or differents parts
}
