% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{tn_bar}
\alias{tn_bar}
\title{theme for barplot}
\usage{
tn_bar(show_axis_title = F, show_legend_title = F, legend_text = 6, ...)
}
\arguments{
\item{show_axis_title}{whether plot axis title}

\item{legend_text}{legend text size}

\item{...}{passing to theme}
}
\value{
theme obj
}
\description{
theme for barplot
}
