% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{tn_italic_text_y}
\alias{tn_italic_text_y}
\title{italic the text y}
\usage{
tn_italic_text_y(...)
}
\arguments{
\item{...}{passing to element_text}
}
\value{
ggplot2 theme
}
\description{
italic the text y
}
