% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{tn_italic_title_x}
\alias{tn_italic_title_x}
\title{italic the title x}
\usage{
tn_italic_title_x(...)
}
\arguments{
\item{...}{passing to element_text}
}
\value{
ggplot2 theme
}
\description{
italic the title x
}
