% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_plot.R
\name{fun_extend_plot_heatmap_cor}
\alias{fun_extend_plot_heatmap_cor}
\title{Plot heatmap with pvalue along the side}
\usage{
fun_extend_plot_heatmap_cor(
  input_df,
  input_x = NULL,
  input_y = data.frame(),
  input_is_cor = F,
  method = "sp",
  fontsize = 5,
  row_names_gp = gpar(cex = 0.5, fontface = "italic"),
  row_split = 1,
  left_col = NULL,
  hm_col = circlize::colorRamp2(breaks = c(-1, 0, 1), colors = c("blue", "white", "red")),
  return_plot_list = F
)
}
\arguments{
\item{input_df}{data.frame with x and rownames of input-y}

\item{input_x}{x}

\item{input_y}{data.frame, rownames should be in input_df}

\item{input_is_cor}{if the results were genrated by \link{fun_stat_cor}}

\item{method}{sp or p}

\item{fontsize}{fontsize}

\item{row_names_gp}{row_names_gp}

\item{row_split}{which data in input_y to split row}

\item{left_col}{left annotation color}

\item{hm_col}{hm_col map}

\item{return_plot_list}{return list}
}
\description{
Plot heatmap with pvalue along the side
}
