% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_roc.R
\name{fun_roc}
\alias{fun_roc}
\title{ROC curve for input data}
\usage{
fun_roc(
  input_df,
  input_variables,
  levels = NULL,
  input_category = "binaryResponse",
  direction = "auto",
  smooth = FALSE
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{levels}{levels}

\item{input_category}{input_category}

\item{direction}{direction}

\item{smooth}{smooth}
}
\value{
list contains roc and auc
}
\description{
ROC curve for input data
}
