% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fisher.R
\name{fun_stat_fisher_by_column}
\alias{fun_stat_fisher_by_column}
\title{Give a data.frame, compare the variable}
\usage{
fun_stat_fisher_by_column(input_df, input_xs, input_ys, echo = F)
}
\arguments{
\item{input_df}{data.frame}

\item{input_xs}{discrete variables}

\item{input_ys}{discrete variables}

\item{echo}{whether print the confusion matrix}
}
\value{
comparison of the fisher results
}
\description{
Give a data.frame, compare the variable
}
