% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fun_train_logit}
\alias{fun_train_logit}
\title{glment train}
\usage{
fun_train_logit(
  input_df,
  input_variables,
  input_y,
  lambda = "lambda.min",
  alpha = 1,
  cv.glment_args = list(),
  seed = 1,
  return_fit = F
)
}
\arguments{
\item{input_df}{input_df}

\item{input_variables}{input_variables}

\item{input_y}{input_y}

\item{lambda}{lambda.min or lambda.1se}

\item{alpha}{0-1}

\item{cv.glment_args}{cv.glment_args}

\item{seed}{seed}

\item{return_fit}{return_fit}
}
\value{
data.frame or list with fit and data.frame
}
\description{
glment train
}
