% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_utils_dt}
\alias{fun_utils_dt}
\title{Using DT to print data.rame}
\usage{
fun_utils_dt(
  input_df,
  input_df_name = "data_download",
  outdir = NULL,
  overwrite = F,
  nrow = 5,
  rownames = F,
  digits = 3,
  fontSize = "75\%"
)
}
\arguments{
\item{input_df}{input_data_frame}

\item{input_df_name}{data_frame_name}

\item{outdir}{outfile to save}

\item{overwrite}{whether to overwrite if file exists}

\item{nrow}{how many rows each page}

\item{rownames}{whether showing the rownames}

\item{digits}{the digits of numeric column}

\item{fontSize}{font size, like "75\%"}
}
\value{
DT obj
}
\description{
Using DT to print data.rame
}
